settings: fix connection addition
Need to make sure we actually export the connection over D-Bus (via
claim_connection()) before we try to return its object path in the
AddConnection reply.  Second, we need to send the path as a string
in the reply, not an object, since the return type is an object path.