settings: don't accept possible auth as success for saving hostname
Since user interaction is allowed when the permission for SaveHostname
is requested, if the user didn't authorize completely when polkit
returns, we don't want to proceed with the request.  Otherwise we might
get into a situation where it's possible for the user to authorize,
but they didn't, and previously the code would allow the request.