keyfile: handle different connections with the same ID
Since keyfile uses the connection's ID as the filename by default,
we could run into a situation where two connections with the same
ID are visible to different users.  We don't want one connection
overwriting the other in that case, so we need to pick a new name
for one of them.  Append the connection's UUID to the end to
minimize the risk of further conflicts for that name.