core/libnm-glib: don't touch files in /etc when updating timestamps
NM updates timestamp for active connections every 5 min. We don't
want to touch files in /etc due to this. This commit solves that
by not updating timestamp in the connection's property. Rather it
stores the timestamp as data associated with connection object. All
timestamps are also kept externally in /var/lib/NetworkManager/timestamps
file.
When settings are requested via D-Bus GetSettings(), the proper
timestamp is put in the setting before they are returned.