build: relax glib version requirement to 2.24
We only need newer glib for G_DEFINE_BOXED_TYPE, which is only
required for introspection.  If you don't want introspection, you
don't need that, so allow building on glib 2.24 without.

We may be compatible with even earlier glib versions too, if so
let me know and I'll downgrade the requirement further.