wimax: handle device removal when the WiMAX Network Service goes away
There's some confusion in the code about indexes; there's the SDK index
from the NS, and there's the internal device index.  Of course we've
got to use the internal device index to remove it from our list, so we
need to convert the SDK index to an internal index before we can call
iwmx_sdk_dev_rm().

Second, the:

if (device_id_list_size < devid->deviceIndex) {

check is useless because by the time we get here, the device is already
gone from the SDK's device list, so this always triggers of the NS
quits.  So just run this check when devices get added to ensure the
NS isn't lying to us.  The, if we get a device removal event, match
up the SDK index with some internal device's SDK index and remove that.