policy: consolidate auto-retries and invalidation
Instead of having two of these, just use the autoconnect retries as
the invalid tag.  Make sure the auto-retries tag always exists by
secretly adding +1 behind the scenes (so the data we set is never NULL).
This should make the code clearer, and will enable the next commit
for authentication failures to decrease the auto retries count.