wifi: fix connection completion when no wifi setting is sent
In NMDeviceWifi's real_complete_connection() the wifi setting
was looked up at the start of the function, but if no wifi
setting was sent by the caller, it would be NULL.  The wifi
setting would later get added by nm_ap_utils_complete_connection(),
but after calling that the new wifi setting would not be looked
up again.  Make that clearer by moving the wifi setting add code
to the wifi device's real_complete_connection() and not burying
it in some other function.  This is more like what other device
types do.