core: aggregate ConsoleKit sesson data (rh #647454)
Jan Schmidt noticed that things didn't work as expected with
multiple sessions of the same user, since when inserting the
new session the old one was forgotten.  Thus bad things happened
if you were local in the old session but not in the new one
since only the new one would be considered.

Instead, make the actual data stored the aggregate of all
sessions for that user.