core: disconnect signals from VPN connection objects when quitting (lp:737713)
A copy & paste error in the VPN connection code meant the VPN
connection was canceling secrets requests for the underlying
device's activation request, not the VPN connection's requests.
Fix that, and as an additional measure, ensure that the manager
removes signal handlers when it gets destroyed that it registered
earlier, preventing those handlers from getting called by the
VPN connection after the manager is dead.  That can happen because
the VPN manager holds references to the VPN connection objects,
and the VPN manager gets destroyed after the NMManager does.

This caused a crash when exiting NM when a VPN connection was active.

(pointed out by dcbw)