build: fix subsequent builds when reconfigured (bgo #648451)
AC_PROG_CXX should be unconditionally run, as it doesn't fail out
if a C++ compiler is not found, which is fine.  It'll check for
one, but we only use it if --enable-qt=yes is given at configure
time.  If --enable-qt=no is set, we do not want to do anything
C++ related, which is why it was conditionalized (incorrectly)
the first time around.